pyproject.toml
[project]
name = "hkex-analysis-agent"
version = "0.1.0"
description = "基于LangGraph的港股公告智能问答系统 - 分层多Agent架构"
authors = [{name = "Development Team"}]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # LangGraph核心
    "langgraph>=0.2.38",
    "langgraph-checkpoint>=2.1.0",
    
    # LangChain
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    
    # LLM集成
    "langchain-siliconflow>=0.1.0",
    "langchain-openai>=0.2.0",
    
    # 数据库
    "clickhouse-connect>=0.7.0",
    
    # Web框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "sse-starlette>=2.0.0",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # 数据验证
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    
    # 配置管理
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    
    # 工具库
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "ipython>=8.0.0",
]

[project.scripts]
hkex-agent = "src.cli.commands:app"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
src = ["**/*.yaml", "**/*.yml"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

